+++ corebehrt/constants/data.py
@@ -15,9 +15,3 @@ TIMESTAMP_COL = "TIMESTAMP"
 BIRTHDATE_COL = "BIRTHDATE"
 DEATHDATE_COL = "DEATHDATE"
 ABSPOS_COL = "abspos"
-
-### Model ###
-TIME2VEC_AGE_MULTIPLIER = 1e-2
-TIME2VEC_ABSPOS_MULTIPLIER = 1e-4
-TIME2VEC_MIN_CLIP = -100
-TIME2VEC_MAX_CLIP = 100
