+++ corebehrt/constants/model.py
@@ -1,0 +1,4 @@
+TIME2VEC_AGE_MULTIPLIER = 1e-2
+TIME2VEC_ABSPOS_MULTIPLIER = 1e-4
+TIME2VEC_MIN_CLIP = -100
+TIME2VEC_MAX_CLIP = 100
