+++ corebehrt/constants/paths.py
@@ -1,0 +1,25 @@
+### Directories ###
+CHECKPOINTS_DIR = "checkpoints"
+
+# Configuration destination names in output folders
+DATA_CFG = "data_config.yaml"
+COHORT_CFG = "cohort_config.yaml"
+OUTCOMES_CFG = "outcomes_config.yaml"
+
+PRETRAIN_CFG = "pretrain_config.yaml"
+FINETUNE_CFG = "finetune_config.yaml"
+
+### Files ###
+VOCABULARY_FILE = "vocabulary.pt"
+PROCESSED_DATA_DIR = "processed_data"
+OUTCOMES_FILE = "outcomes.csv"
+INDEX_DATES_FILE = "index_dates.csv"
+PID_FILE = "pids.pt"
+
+### File Patterns ###
+CONCEPT_FORMAT = "concept.*"
+PATIENTS_INFO_FORMAT = "patients_info.*"
+
+### File Extensions ###
+CSV_EXT = ".csv"
+PARQUET_EXT = ".parquet"
