+++ corebehrt/functional/features/tokenize.py
@@ -1,10 +1,8 @@
 import pandas as pd
-from corebehrt.functional.constants import (
-    SPECIAL_TOKEN_ABSPOS_ADJUSTMENT,
-    UNKNOWN_TOKEN,
-    CLS_TOKEN,
-    SEP_TOKEN,
-)
+
+from corebehrt.constants.data import (CLS_TOKEN, SEP_TOKEN,
+                                      SPECIAL_TOKEN_ABSPOS_ADJUSTMENT,
+                                      UNKNOWN_TOKEN)
 
 
 def add_special_tokens_partition(
