+++ corebehrt/modules/features/normalizer.py
@@ -1,4 +1,5 @@
 from dask import dataframe as dd
+
 from corebehrt.functional.features.normalize import min_max_normalize
