+++ corebehrt/modules/preparation/prepare_data.py
@@ -9,8 +9,8 @@ import pandas as pd
 import torch
 from dask.diagnostics import ProgressBar
 
+from corebehrt.constants.data import ABSPOS_COL, PID_COL, TIMESTAMP_COL
 from corebehrt.functional.cohort_handling.outcomes import get_binary_outcomes
-from corebehrt.functional.constants import ABSPOS_COL, PID_COL, TIMESTAMP_COL
 from corebehrt.functional.features.normalize import normalize_segments_for_patient
 from corebehrt.functional.io_operations.load import load_vocabulary
 from corebehrt.functional.io_operations.save import save_vocabulary
@@ -28,7 +28,7 @@ from corebehrt.functional.preparation.utils import (
 from corebehrt.modules.cohort_handling.patient_filter import filter_df_by_pids
 from corebehrt.modules.preparation.dataset import PatientDataset
 from corebehrt.modules.setup.config import Config, load_config
-from corebehrt.modules.setup.directory import (
+from corebehrt.constants.paths import (
     INDEX_DATES_FILE,
     OUTCOMES_FILE,
     PID_FILE,
