+++ tests/test_functional/test_outcomes.py
@@ -1,7 +1,7 @@
 import unittest
 import pandas as pd
 
-from corebehrt.functional.constants import PID_COL, ABSPOS_COL
+from corebehrt.constants.data import PID_COL, ABSPOS_COL
 from corebehrt.functional.cohort_handling.outcomes import get_binary_outcomes
