+++ tests/test_modules/test_index_date_handler.py
@@ -3,8 +3,8 @@ from datetime import datetime
 
 import pandas as pd
 
+from corebehrt.constants.data import PID_COL, TIMESTAMP_COL
 from corebehrt.modules.cohort_handling.index_dates import IndexDateHandler
-from corebehrt.functional.constants import PID_COL, TIMESTAMP_COL
 
 
 class TestIndexDateHandler(unittest.TestCase):
